% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestcox.R
\name{count_event_by}
\alias{count_event_by}
\title{FUNCTION_TITLE}
\usage{
count_event_by(
  formula,
  data,
  count_by_var = NULL,
  var_subgroup = NULL,
  decimal.percent = 1
)
}
\arguments{
\item{formula}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION}

\item{count_by_var}{PARAM_DESCRIPTION, Default: NULL}

\item{var_subgroup}{PARAM_DESCRIPTION, Default: NULL}

\item{decimal.percent}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tidyr]{drop_na}}
 \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{arrange}}
}
