% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxme.R
\name{coxExp}
\alias{coxExp}
\title{coxExp: transform the unit of coefficients in cox model(internal function)}
\usage{
coxExp(cox.coef, dec)
}
\arguments{
\item{cox.coef}{cox model coefficients}

\item{dec}{Decimal point}
}
\value{
The transforemed coefficients(95% CI), p-value
}
\description{
Transform the unit of coefficients to "HR"
}
\details{
DETAILS
}
\examples{
 library(coxme)
 fit <- coxme(Surv(time, status) ~ ph.ecog + age + (1|inst), lung)
 jstable:::coxExp(jstable:::coxmeTable(fit))
}
