% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxme.R
\name{coxmeTable}
\alias{coxmeTable}
\title{coxmeTable: Summary table of coxme.object(internal function)}
\usage{
coxmeTable(mod)
}
\arguments{
\item{mod}{coxme.object}
}
\value{
beta, se, z, p of fixed effects
}
\description{
Extract fixed effect table in coxme.object
}
\details{
DETAILS
}
\examples{
library(coxme)
fit <- coxme(Surv(time, status) ~ ph.ecog + age + (1 | inst), lung)
jstable:::coxmeTable(fit)
}
