% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{geeglm.display}
\alias{geeglm.display}
\title{geeglm.display}
\usage{
geeglm.display(geeglm.obj, decimal = 2)
}
\arguments{
\item{geeglm.obj}{"geeglm" object}

\item{decimal}{Decimal, Default: 2}
}
\value{
List: caption, main table, metrics table
}
\description{
Make gee results from "geeglm" object
}
\details{
DETAILS
}
\examples{
 library(geepack)
 data(dietox)
 dietox$Cu <- as.factor(dietox$Cu)
 gee01 <- geeglm (Weight ~ Time + Cu , id =Pig, data = dietox,
                family=gaussian,corstr="ex")
 geeglm.display(gee01)
}
\seealso{
\code{\link[data.table]{data.table-package}}
 \code{\link[stats]{complete.cases}}
}
