% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmshow.R
\name{glmshow.display}
\alias{glmshow.display}
\title{glmshow.display: Show summary table of glm object.}
\usage{
glmshow.display(glm.object, decimal = 2)
}
\arguments{
\item{glm.object}{glm.object}

\item{decimal}{digits, Default: 2}
}
\value{
table
}
\description{
Show summary table of glm object(regression, logistic).
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{glm}}
}
