% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsBasicAddin.R
\name{jsBasicGadget}
\alias{jsBasicGadget}
\title{jsBasicGadget: Shiny Gadget of Basic Statistics in Medical Research.}
\usage{
jsBasicGadget(data)
}
\arguments{
\item{data}{data}
}
\value{
Shiny Gadget including Data, Label info, Table 1, Regression(liear, logistic), Basic plot
}
\description{
Shiny Gadget including Data, Label info, Table 1, Regression(liear, logistic), Basic plot
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 library(survival)
 jsBasicGadget(lung) 
 }
}
}
