% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{LabeljsGeeglm}
\alias{LabeljsGeeglm}
\title{LabeljsGeeglm: Apply label information to geeglm.display object using label data}
\usage{
LabeljsGeeglm(obj, ref)
}
\arguments{
\item{obj}{geeglm.display object}

\item{ref}{Label data made by mk.lev function}
}
\value{
geeglm.display object with label information
}
\description{
Apply label information to geeglm.display object using label data
}
\details{
DETAILS
}
\examples{
library(geepack)
library(jstable)
data(dietox)
dietox$Cu <- as.factor(dietox$Cu)
gee01 <- geeglm(Weight ~ Time + Cu,
  id = Pig, data = dietox,
  family = gaussian, corstr = "ex"
)
g1 <- geeglm.display(gee01)
LabeljsGeeglm(g1, ref = mk.lev(dietox))
}
