% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{LabeljsMixed}
\alias{LabeljsMixed}
\title{LabeljsMixed: Apply label information to jstable object using label data}
\usage{
LabeljsMixed(obj, ref)
}
\arguments{
\item{obj}{lmer.display, coxme.display}

\item{ref}{Label data made by mk.lev function}
}
\value{
lmer.display, coxme.display with label information
}
\description{
Apply label information to object of jstable using label data
}
\details{
DETAILS
}
\examples{
library(coxme)
fit <- coxme(Surv(time, status) ~ sex + ph.ecog + ph.karno + (1 | inst) + (1 | sex), lung)
fit.table <- coxme.display(fit)
lung.label <- mk.lev(lung)
LabeljsMixed(fit.table, ref = lung.label)
}
