% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer.R
\name{lmerExp}
\alias{lmerExp}
\title{lmerExp: transform the unit of coefficients (internal function)}
\usage{
lmerExp(lmer.coef, family = "binomial", dec)
}
\arguments{
\item{lmer.coef}{lmer coefficients.}

\item{family}{Family: "gaussian", "binomial", "poisson", "quasipoisson", etc..., Default: 'binomial'}

\item{dec}{Decimal point}
}
\value{
The transforemed coefficients(95% CI), p-value
}
\description{
Transform the unit of coefficients to "Coeff", "OR" or "RR"
}
\details{
DETAILS
}
\examples{
# EXAMPLE1
}
