% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{mk.lev}
\alias{mk.lev}
\title{Export label and level: multiple variable}
\usage{
mk.lev(data)
}
\arguments{
\item{data}{data}
}
\value{
default label and level data
}
\description{
Export label and level: multiple variable
}
\details{
DETAILS
}
\examples{
 mk.lev(iris)
}
