% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{mk.lev.var}
\alias{mk.lev.var}
\title{Export label and level: one variable}
\usage{
mk.lev.var(data, vname)
}
\arguments{
\item{data}{data}

\item{vname}{variable to export label and level}
}
\value{
if continuous variable - (label, NA), categorical variable - (label, level)
}
\description{
Export label and level: one variable
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 out.list = lapply(names(data), function(x){mk.lev.var(data, x)})
 }
}
}
