% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmshow.R
\docType{data}
\name{mort}
\alias{mort}
\title{DATASET_TITLE}
\format{
A data frame with 17562 rows and 24 variables:
\describe{
  \item{\code{ccode}}{integer COLUMN_DESCRIPTION}
  \item{\code{cname}}{character COLUMN_DESCRIPTION}
  \item{\code{yy}}{integer COLUMN_DESCRIPTION}
  \item{\code{mm}}{integer COLUMN_DESCRIPTION}
  \item{\code{dd}}{integer COLUMN_DESCRIPTION}
  \item{\code{date}}{character COLUMN_DESCRIPTION}
  \item{\code{nonacc}}{integer COLUMN_DESCRIPTION}
  \item{\code{cardio}}{integer COLUMN_DESCRIPTION}
  \item{\code{respir}}{integer COLUMN_DESCRIPTION}
  \item{\code{influenza}}{integer COLUMN_DESCRIPTION}
  \item{\code{meanpm10}}{double COLUMN_DESCRIPTION}
  \item{\code{meanso2}}{double COLUMN_DESCRIPTION}
  \item{\code{meanno2}}{double COLUMN_DESCRIPTION}
  \item{\code{meanco}}{double COLUMN_DESCRIPTION}
  \item{\code{maxco}}{double COLUMN_DESCRIPTION}
  \item{\code{maxo3}}{double COLUMN_DESCRIPTION}
  \item{\code{meantemp}}{double COLUMN_DESCRIPTION}
  \item{\code{maxtemp}}{double COLUMN_DESCRIPTION}
  \item{\code{mintemp}}{double COLUMN_DESCRIPTION}
  \item{\code{meanhumi}}{double COLUMN_DESCRIPTION}
  \item{\code{meanpress}}{double COLUMN_DESCRIPTION}
  \item{\code{season}}{integer COLUMN_DESCRIPTION}
  \item{\code{dow}}{integer COLUMN_DESCRIPTION}
  \item{\code{sn}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
mort
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
