% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DToption.R
\name{opt.data}
\alias{opt.data}
\title{datable option for data(DT package)}
\usage{
opt.data(fname)
}
\arguments{
\item{fname}{File name to download}
}
\value{
datatable object
}
\description{
DT::datatable option for data
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 datatable(a.label, rownames=F, extension= "Buttons", caption = "Labels of data",
 options = opt.data("data")
 )
 }
}
}
