% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DToption.R
\name{opt.roc}
\alias{opt.roc}
\title{datable option for ROC result(DT package)}
\usage{
opt.roc(fname)
}
\arguments{
\item{fname}{File name to download}
}
\value{
datatable option object
}
\description{
DT::datatable option for ROC result
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 datatable(roc.table, rownames=F, extension= "Buttons", caption = "Best cut-off", 
           options = opt.roc(paste("roc", input$dep_vars_roc , 
                                   paste(xs = input$indep_vars_roc, collapse = "_"), sep="_"))
           )
 }
}
}
