% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DToption.R
\name{opt.tb1}
\alias{opt.tb1}
\title{datable option for table 1(DT package)}
\usage{
opt.tb1(fname)
}
\arguments{
\item{fname}{File name to download}
}
\value{
datatable object
}
\description{
DT::datatable option for table 1
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 datatable(ptb1, rownames=T, extension= "Buttons", caption = cap.tb1,
 options = c(opt.tb1("tb1"), list(columnDefs = list(list(visible=FALSE, targets=ncol(ptb1)))
                                 )
            )
 }
}
}
