% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DToption.R
\name{opt.tbreg}
\alias{opt.tbreg}
\title{datable option for regression table(DT package)}
\usage{
opt.tbreg(fname)
}
\arguments{
\item{fname}{File name to download}
}
\value{
datatable object
}
\description{
DT::datatable option for glm, gee(geepack packge), lmer/glmer(lme4 package)
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 datatable(out.gee, rownames=T, extension= "Buttons", caption = cap.gee,
          options = c(opt.tbreg(paste("gee_gaussian", y , paste(xs, collapse = "_"), sep="_"))
                     )
          )
 }
}
}
