% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyglm.R
\name{svyglm.display}
\alias{svyglm.display}
\title{svyglm.display: table for svyglm.object}
\usage{
\method{svyglm}{display}(svyglm.obj, decimal = 2)
}
\arguments{
\item{svyglm.obj}{svyglm.object}

\item{decimal}{digit, Default: 2}
}
\value{
table
}
\description{
table for svyglm.object (survey package).
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 data(api)
 apistrat$tt = c(rep(1, 20), rep(0, nrow(apistrat) -20))
 dstrat<-svydesign(id=~1,strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
 ds <- svyglm(api00~ell+meals+cname+mobility, design=dstrat)
 ds2 <- svyglm(tt~ell+meals+cname+mobility, design=dstrat, family = quasibinomial())
 svyglm.display(ds)
 svyglm.display(ds2)
 }
}
}
