% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_event_by}
\alias{count_event_by}
\title{count_event_by: funciton to count event, subgroup number inside TableSubgroupCox, TableSubgroupMultiCox}
\usage{
count_event_by(
  formula,
  data,
  count_by_var = NULL,
  var_subgroup = NULL,
  decimal.percent = 1
)
}
\arguments{
\item{formula}{formula with survival analysis}

\item{data}{same data as in formula}

\item{count_by_var}{variables to count subgroup for}

\item{var_subgroup}{1 sub-group variable for analysis,}

\item{decimal.percent}{decimals to show percent of, Default: 1}
}
\value{
Table with event, subgroup number
}
\description{
Function to count event, subgroup number
}
\details{
This function is used inside TableSubgroupCox, TableSubgroupMultiCox for calculation
}
\examples{
\dontrun{
if (interactive()) {

}
}
}
\seealso{
\code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{arrange}}
}
