% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DToption.R
\name{opt.simpledown}
\alias{opt.simpledown}
\title{datable option for simple download(DT package)}
\usage{
opt.simpledown(fname)
}
\arguments{
\item{fname}{File name to download}
}
\value{
datatable option object
}
\description{
Simple download DT::datatable option - No filter, No page
}
\details{
DETAILS
}
\examples{
 DT::datatable(mtcars, rownames = FALSE, extension= "Buttons", caption = "mtcars",
 options = opt.simpledown("mtcars"))
}
