% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DToption.R
\name{opt.tb1}
\alias{opt.tb1}
\title{datable option for table 1(DT package)}
\usage{
opt.tb1(fname)
}
\arguments{
\item{fname}{File name to download}
}
\value{
datatable option object
}
\description{
DT::datatable option for table 1
}
\details{
DETAILS
}
\examples{
 DT::datatable(mtcars, rownames = FALSE, extension= "Buttons", caption = "mtcars",
 options = opt.tb1("mtcars"))
}
