% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DToption.R
\name{opt.tbreg}
\alias{opt.tbreg}
\title{datable option for regression table(DT package)}
\usage{
opt.tbreg(fname)
}
\arguments{
\item{fname}{File name to download}
}
\value{
datatable option object
}
\description{
DT::datatable option for glm, gee(geepack package), lmer/glmer(lme4 package)
}
\details{
DETAILS
}
\examples{
 options = opt.tbreg("mtcars")
}
