% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyglm.R
\name{svyregress.display}
\alias{svyregress.display}
\title{svyregress.display: table for svyglm.object}
\usage{
svyregress.display(svyglm.obj, decimal = 2)
}
\arguments{
\item{svyglm.obj}{svyglm.object}

\item{decimal}{digit, Default: 2}
}
\value{
table
}
\description{
table for svyglm.object (survey package).
}
\details{
DETAILS
}
\examples{
 library(survey);data(api)
 apistrat$tt = c(rep(1, 20), rep(0, nrow(apistrat) -20))
 dstrat<-svydesign(id=~1,strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
 ds <- svyglm(api00~ell+meals+cname+mobility, design=dstrat)
 ds2 <- svyglm(tt~ell+meals+cname+mobility, design=dstrat, family = quasibinomial())
 svyregress.display(ds)
 svyregress.display(ds2)
}
